backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
Pillow
python-multipart
